requirements.txt
fastapi[standard]>=0.116
langgraph>=0.2
pydantic>=2.7
python-dotenv>=1.0
requests>=2.32
